requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
gensim==4.1.2
idna==3.3
joblib==1.1.0
newsapi-python==0.2.6
nltk==3.6.5
numpy==1.21.4
regex==2021.11.10
requests==2.26.0
scipy==1.7.3
smart-open==5.2.1
tqdm==4.62.3
typer==0.4.0
urllib3==1.26.7
